backend/requirements.txt
Flask
Flask-PyMongo
Flask-Cors
python-dotenv
pymongo
Flask-Mail
